+++ vm/src/utils.rs
@@ -256,13 +256,13 @@ pub mod test_utils {
                 false,
                 false,
                 false,
-                false
+                false,
             )
             .unwrap()
         };
         ($program:expr, $layout:expr) => {
             crate::vm::runners::cairo_runner::CairoRunner::new(
-                &$program, $layout, None, false, false, false, false
+                &$program, $layout, None, false, false, false, false,
             )
             .unwrap()
         };
@@ -274,7 +274,7 @@ pub mod test_utils {
                 $proof_mode,
                 false,
                 false,
-                false
+                false,
             )
             .unwrap()
         };
@@ -286,7 +286,7 @@ pub mod test_utils {
                 $proof_mode,
                 $trace_enabled,
                 false,
-                false
+                false,
             )
             .unwrap()
         };