+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 #### Upcoming Changes
 
+* feat(BREAKING): add allow_disordered_builtins to CairoRunner to avoid checking builtins order. [#2292](https://github.com/lambdaclass/cairo-vm/pull/2292)
+
 * fix: Make cairo1-run to conditionally relocate memory and trace [#2241](https://github.com/lambdaclass/cairo-vm/pull/2241)
 
 #### [3.0.1] - 2025-12-22